#38 - Secuity Casting 에러 해결 (casting -> getName())
