+++ src/main/java/com/artfolio/artfolio/business/controller/WebSocketController.java
@@ -4,7 +4,7 @@ import com.artfolio.artfolio.business.dto.AuctionBid;
 import com.artfolio.artfolio.global.error.ErrorCode;
 import com.artfolio.artfolio.global.error.ErrorResponse;
 import com.artfolio.artfolio.global.exception.InvalidBidPriceException;
-import com.artfolio.artfolio.business.service.RealTimeAuctionService;
+import com.artfolio.artfolio.business.service.AuctionService;
 import com.artfolio.artfolio.global.exception.UserNotFoundException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -13,6 +13,7 @@ import org.springframework.messaging.handler.annotation.MessageExceptionHandler;
 import org.springframework.messaging.handler.annotation.MessageMapping;
 import org.springframework.messaging.handler.annotation.Payload;
 import org.springframework.messaging.simp.SimpMessageSendingOperations;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -25,13 +26,13 @@ import static com.artfolio.artfolio.global.util.ErrorBuildFactory.buildError;
 @RestController
 public class WebSocketController {
     private final SimpMessageSendingOperations simp;
-    private final RealTimeAuctionService realTimeAuctionService;
+    private final AuctionService realTimeAuctionService;
 
     /* 가격 실시간 갱신 */
     // 구독 경로 : /sub/channel/{auctionId}
     // 발행 경로 : /pub/price
     @MessageMapping("/price")
-    public void updateAuctionPrice(Principal principal, @Payload AuctionBid.Req req) {
+    public void updateAuctionPrice(@AuthenticationPrincipal Principal principal, @Payload AuctionBid.Req req) {
         AuctionBid.Res res = realTimeAuctionService.updatePrice(principal, req);
         simp.convertAndSend("/topic/channel/" + req.getAuctionId(), res);
     }
