+++ src/main/java/com/artfolio/artfolio/business/domain/ArtPiece.java
@@ -34,7 +34,7 @@ public class ArtPiece extends AuditingFields {
     private final List<ArtPiecePhoto> artPiecePhotos = new ArrayList<>();
 
     @Column(nullable = false)
-    @OneToMany(mappedBy = "artPiece")
+    @OneToMany(mappedBy = "artPiece", fetch = FetchType.EAGER)
     private final List<Auction> auctions = new ArrayList<>();
 
     @Builder
@@ -52,5 +52,10 @@ public class ArtPiece extends AuditingFields {
                 .like(0L)
                 .artist(artist)
                 .build();
+    }
+
+    public void updatePhoto(ArtPiecePhoto artPiecePhoto) {
+        this.artPiecePhotos.add(artPiecePhoto);
+        artPiecePhoto.setArtPiece(this);
     }
 }
