--- src/main/java/com/artfolio/artfolio/business/dto/AuctionDetails.java
@@ -1,19 +1,0 @@
-package com.artfolio.artfolio.business.dto;
-
-import com.artfolio.artfolio.business.domain.ArtPiecePhoto;
-import com.artfolio.artfolio.user.entity.User;
-import com.artfolio.artfolio.business.domain.redis.AuctionBidInfo;
-import com.artfolio.artfolio.business.domain.redis.RealTimeAuctionInfo;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-
-import java.time.LocalDateTime;
-import java.util.Comparator;
-import java.util.List;
-import java.util.Set;
-
-
-public class AuctionDetails {
-
-}
