+++ src/main/java/com/artfolio/artfolio/business/service/ImageService.java
@@ -26,11 +26,10 @@ public class ImageService {
             + "/src/main/resources/images";
     private final ArtPiecePhotoRepository artPiecePhotoRepository;
     private final ArtPieceRepository artPieceRepository;
-    private final RealTimeAuctionService realTimeAuctionService;
+    private final AuctionService realTimeAuctionService;
     private final S3Uploader s3Uploader;
     private String thumbnailFileName;
 
-    /*
     public Long uploadImage(Long artPieceId, MultipartFile[] files) {
         ArtPiece artPiece = artPieceRepository.findById(artPieceId)
                 .orElseThrow(() -> new ArtPieceNotFoundException(artPieceId));
@@ -60,9 +59,8 @@ public class ImageService {
                         .build();
 
                 photos.add(entity);
+                artPiece.updatePhoto(entity);
             }
-
-            realTimeAuctionService.updateImage(artPieceId, s3Path);
         }
 
         // 로컬 경로 내 모든 사진 삭제
@@ -74,7 +72,6 @@ public class ImageService {
 
         return 1L;
     }
-    */
 
     /* resources/images 경로에 원본 이미지와 압축된 이미지를 생성해주는 메서드 */
     private boolean saveImageOnLocal(MultipartFile[] files, File imgDir) {
