+++ src/main/java/com/artfolio/artfolio/global/config/AuditingConfig.java
@@ -14,9 +14,10 @@ public class AuditingConfig {
     @Bean
     public AuditorAware<String> auditorAware() {
         return () -> Optional.of(
-                (String) SecurityContextHolder
+                SecurityContextHolder
                         .getContext()
                         .getAuthentication()
-                        .getPrincipal());
+                        .getName()
+        );
     }
 }
